Improve toggle switch UX, add hover styles
This commit gets rid of the "dead" zone in the toggle switch, i.e., the
area of the toggle button. Previously, clicking on the active area of
the toggle switch doesn't "toggle" between the two options, since these
are just radio buttons and clicking on an already active radio button
does nothing.

To solve this, we make the inactive radio button span the width of the
toggle switch and place it on top, so that clicking anywhere within the
component will switch the active option just like a true toggle switch.

Apart from that, this doesn't break keyboard accessibility.

Clicking on a label won't toggle the switch; it will just select the
corresponding option. This provides users with an explicit way to set
the billing method. Besides, it would be weird to click on the Monthly
label and have the billing method set to Annually.